rand_os backward compatibility
