+++ rand_os/src/impls/cloudabi.rs
