+++ rand_os/src/impls/dragonfly_haiku_emscripten.rs
