+++ rand_os/src/impls/dummy_log.rs
