+++ rand_os/src/impls/fuchsia.rs
