+++ rand_os/src/impls/linux_android.rs
