+++ rand_os/src/impls/netbsd.rs
