+++ rand_os/src/impls/openbsd_bitrig.rs
