+++ rand_os/src/impls/redox.rs
