+++ rand_os/src/impls/solaris.rs
