+++ rand_os/src/impls/wasm32_bindgen.rs
