+++ rand_os/src/impls/wasm32_stdweb.rs
